schwab-equities-workstation/backend/pyproject.toml
[project]
name = "schwab-equities-workstation-backend"
version = "0.1.0"
description = "Backend API for a Schwab-ready equities analysis workstation"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.6",
  "pydantic>=2.9.0",
  "schwab-py>=1.5.0",
  "python-dotenv>=1.0.1"
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
